Adds navigation and separates into individual files
